Refactoring the code to use the model classes.
Now we remove all that logic that was mixed with html from our files.
The logic is now stored in our models.
With this we get a better organization in our code.
If we need to change the rules of our project, we need only to look at the model files.