Real views and controllers now.
A Controller base class was created for giving more power to our
controllers.
Now the controllers have the ability to `render` a view passing it an
array of variables if needed.
With this change, our views are now only responsible for the presentation
of our web site info.
The front controller was also adapted to select the apropriate controllers
instead of views for our routes.