Adding a router for more organization.
We created a Router class to handle the routes and choose the apropriate
controller and action.
So now we remove or routes from the index.php and put them on a specific
config file.
And the index.php now is much cleaner. We only need to define the
baseUrl in it depending on our environment.